+++ src/sao_modules/snake_module/headers.hpp
@@ -40,9 +40,9 @@ public:
 
     void setup();
 
-    void displayUpdate(int64_t lastMetaDisplayUpdate);
+    void displayUpdate(uint32_t lastMetaDisplayUpdate);
     
-    void logicUpdate(int64_t lastMetaLogicUpdate);
+    void logicUpdate(uint32_t lastMetaLogicUpdate);
     
     void teardown();
